#602: Sonar fixes.

+++ lionengine-game/src/main/java/com/b3dgs/lionengine/game/feature/tile/TileGame.java
@@ -77,8 +77,8 @@ public class TileGame implements Tile
         this.height = height;
 
         key = Integer.valueOf(number);
-        x = tx * width;
-        y = ty * height;
+        x = tx * (double) width;
+        y = ty * (double) height;
     }
 
     /**
